pyproject.toml
[project]
name = "phd-hunter"
version = "0.1.0"
description = "PhD 导师套磁筛选助手 - Automated PhD advisor matching and analysis system"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "PhD Hunter Team", email = "team@phdhunter.dev" }
]
keywords = ["phd", "academic", "research", "advisor", "csrankings", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "requests>=2.31.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.14.0",
    "selenium>=4.20.0",
    "webdriver-manager>=4.0.0",
    "lxml>=6.0.0",
    "arxiv>=2.0.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "pydantic>=2.7.0",
    "loguru>=0.7.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "streamlit>=1.35.0",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "tenacity>=9.0.0",
    "tqdm>=4.66.0",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=3.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "pytest-asyncio>=0.23.0",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # Line too long (handled by black)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=phd_hunter --cov-report=term-missing"

[tool.coverage.run]
source = ["phd_hunter"]
omit = [
    "tests/*",
    "frontend/*",
    "**/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
